Move gwas classes to it's plugin.
Also remove SearchKeywordService duplicate
