+++ grails-app/services/org/transmart/SearchKeywordService.groovy
@@ -1,29 +1,31 @@
 package org.transmart
+
+import org.transmart.biomart.BioAssayPlatform
+import org.transmart.biomart.BioDataExternalCode
+import org.transmart.biomart.ConceptCode
+import org.transmart.searchapp.AuthUser
+import org.transmart.searchapp.GeneSignature
+
 /*************************************************************************
  * tranSMART - translational medicine data mart
- * 
+ *
  * Copyright 2008-2012 Janssen Research & Development, LLC.
- * 
+ *
  * This product includes software developed at Janssen Research & Development, LLC.
- * 
+ *
  * This program is free software: you can redistribute it and/or modify it under the terms of the GNU General Public License 
  * as published by the Free Software  * Foundation, either version 3 of the License, or (at your option) any later version, along with the following terms:
  * 1.	You may convey a work based on this program in accordance with section 5, provided that you retain the above notices.
  * 2.	You may convey verbatim copies of this program code as you receive it, in any medium, provided that you retain the above notices.
- * 
+ *
  * This program is distributed in the hope that it will be useful, but WITHOUT ANY WARRANTY; without even the implied warranty of MERCHANTABILITY or FITNESS    * FOR A PARTICULAR PURPOSE.  See the GNU General Public License for more details.
- * 
+ *
  * You should have received a copy of the GNU General Public License along with this program.  If not, see http://www.gnu.org/licenses/.
- * 
+ *
  *
  ******************************************************************/
-  
-import org.transmart.searchapp.AuthUser;
-
-
 import org.transmart.searchapp.SearchKeyword
 import org.transmart.searchapp.SearchKeywordTerm
-import org.transmart.searchapp.GeneSignature
 
 /**
  * @author $Author: mmcduffie $
@@ -33,353 +35,496 @@ import org.transmart.searchapp.GeneSignature
  */
 public class SearchKeywordService {
 
-	def springSecurityService
-	
-	// probably not needed but makes all methods transactional
-	static transactional = true
+    def springSecurityService
 
-	/** Finds all of the search categories pertaining to search keywords */
-	def findSearchCategories()	{
-		log.info "Finding all of the search categories..."
-		
-		def c = SearchKeyword.createCriteria()
-		def results = c.list	{
-			projections {
-				distinct("dataCategory")
-			}
-			order("dataCategory", "asc")
-		}
-		
-		log.info("Categories found: " + results.size())
-		
-		def categories = []
-		
-		for (result in results)	{
-			categories.add(["category":result])
-		}
-		
-		return categories
-	}
-	
-	/** Searches for all keywords for a given term (like %il%) */
-	def findSearchKeywords(category, term, max)	{
-		log.info "Finding matches for ${term} in ${category}"
-		
-		def user = org.transmart.searchapp.AuthUser.findByUsername(springSecurityService.getPrincipal().username)
-		
-		def c = SearchKeywordTerm.createCriteria()
-		def results = c.list 	{
-			if (term.size() > 0)	{
-				like("keywordTerm", term.toUpperCase() + '%')
-			}
-			
-			//TODO Special case for gene or SNP - rework to support multiple categories!
-			if ("GENE_OR_SNP".equals(category))	{
-				//or {
-				//	eq("dataCategory", "GENE")
-				//	like("keywordTerm", 'RS%')
-				//}
-				'in'("dataCategory",["GENE","SNP"])
-			}
-			else if ("SNP".equals(category)) {
-				//like("keywordTerm", 'RS%')
-				eq("dataCategory", "SNP")
-			}
-			else if ("ALL".compareToIgnoreCase(category) != 0)	{
-				eq("dataCategory", category.toUpperCase())
-			}
+    // probably not needed but makes all methods transactional
+    static transactional = true
 
-			if (!user.isAdmin())	{
-				log.info("User is not an admin so filter out gene lists or signatures that are not public")
-				or	{
-					isNull("ownerAuthUserId")
-					eq("ownerAuthUserId", user.id)
-				}
-			}
-			maxResults(max)
-			order("rank", "asc")
-			order("termLength", "asc")
-			order("keywordTerm", "asc")
-		}
-		log.info("Search keywords found: " + results.size())
-		
-		def keywords = []
-		def dupeList = []			// store category:keyword for a duplicate check until DB is cleaned up
-		
-		for (result in results)	{
-			def m = [:]
-			def sk = result
-			//////////////////////////////////////////////////////////////////////////////////
-			// HACK:  Duplicate check until DB is cleaned up
-			def dupeKey = sk.searchKeyword.displayDataCategory + ":" +sk.searchKeyword.keyword
-			if (dupeKey in dupeList)	{
-				log.info "Found duplicate: " + dupeKey
-				continue
-			} else	{
-				log.info "Found new entry, adding to the list: " + dupeList
-				dupeList << dupeKey
-			}
-			///////////////////////////////////////////////////////////////////////////////////
-			
-			m.put("label", sk.searchKeyword.keyword)
-			m.put("category", sk.searchKeyword.displayDataCategory)
-			
-			
-			//Further hack: Alter fields depending on the category
-			if (sk.searchKeyword.dataCategory.equals("DISEASE") || sk.searchKeyword.dataCategory.equals("OBSERVATION")) {
-				m.put("categoryId", sk.searchKeyword.dataCategory)
-				m.put("id", sk.searchKeyword.uniqueId)
-			}
-			else {
-				m.put("categoryId", sk.searchKeyword.dataCategory)
-				m.put("id", sk.searchKeyword.id)
-			}
-			if ("TEXT".compareToIgnoreCase(sk.searchKeyword.dataCategory) != 0)	{
-				def synonyms = org.transmart.biomart.BioDataExternalCode.findAllWhere(bioDataId: sk.searchKeyword.bioDataId, codeType: "SYNONYM")
-				def synList = new StringBuilder()
-				for (synonym in synonyms)	{
-					if (synList.size() > 0)	{
-						synList.append(", ")
-					} else	{
-						synList.append("(")
-					}
-					synList.append(synonym.code)
-				}
-				if (synList.size() > 0)	{
-					synList.append(")")
-				}
-				m.put("synonyms", synList.toString())
-			}
-			keywords.add(m)
-		}
-		return keywords
-	 }
+    //Hard-coded list of items that we consider filter categories... configure in Config/database?
+    def filtercats = [
+
+            [codeTypeName: "THERAPEUTIC_DOMAIN", category: "THERAPEUTIC_DOMAIN", displayName: "Program Therapeutic Domain"],
+            [codeTypeName: "PROGRAM_INSTITUTION", category: "PROGRAM_INSTITUTION", displayName: "Program Institution"],
+
+            [codeTypeName: "STUDY_PHASE", category: "STUDY_PHASE", displayName: "Study Phase"],
+            [codeTypeName: "STUDY_OBJECTIVE", category: "STUDY_OBJECTIVE", displayName: "Study Objective"],
+            [codeTypeName: "STUDY_DESIGN", category: "STUDY_DESIGN", displayName: "Study Design"],
+            [codeTypeName: "STUDY_BIOMARKER_TYPE", category: "STUDY_BIOMARKER_TYPE", displayName: "Study Biomarker Type"],
+            [codeTypeName: "STUDY_ACCESS_TYPE", category: "STUDY_ACCESS_TYPE", displayName: "Study Access Type"],
+            [codeTypeName: "STUDY_INSTITUTION", category: "STUDY_INSTITUTION", displayName: "Study Institution"],
+
+            [codeTypeName: "ASSAY_TYPE_OF_BM_STUDIED", category: "ASSAY_TYPE_OF_BM_STUDIED", displayName: "Assay Type of Biomarkers"],
+            [category: "ASSAY_MEASUREMENT_TYPE", displayName: "Assay Measurement Type", useText: true, platformProperty: 'platformType'],
+            [category: "ASSAY_TECHNOLOGY", displayName: "Assay Technology", prefix: true, useText: true, platformProperty: 'platformTechnology'],
+            [category: "ASSAY_VENDOR", displayName: "Assay Vendor", prefix: true, useText: true, platformProperty: 'vendor'],
+            [category: "ASSAY_PLATFORM_NAME", displayName: "Assay Platform Name", useText: true, platformProperty: 'name'],
+
+            [category: "ANALYSIS_MEASUREMENT_TYPE", displayName: "Analysis Measurement Type", useText: true, platformProperty: 'platformType'],
+            [category: "ANALYSIS_TECHNOLOGY", displayName: "Analysis Technology", prefix: true, useText: true, platformProperty: 'platformTechnology'],
+            [category: "ANALYSIS_VENDOR", displayName: "Analysis Vendor", prefix: true, useText: true, platformProperty: 'vendor'],
+            [category: "ANALYSIS_PLATFORM_NAME", displayName: "Analysis Platform Name", useText: true, platformProperty: 'name'],
+
+            [codeTypeName: "FILE_TYPE", category: "FILE_TYPE", displayName: "File type"]
+    ]
+
+    /** Finds all of the search categories pertaining to search keywords */
+    def findSearchCategories() {
+        log.info "Finding all of the search categories..."
+
+        def c = SearchKeyword.createCriteria()
+        def results = c.list {
+            projections {
+                distinct("dataCategory")
+            }
+            order("dataCategory", "asc")
+        }
+
+        log.info("Categories found: " + results.size())
+
+        def categories = []
+
+        for (result in results) {
+            categories.add(["category": result])
+        }
+
+        return categories
+    }
+
+    def findFilterCategories() {
+
+        def categories = []
+
+        for (filtercat in filtercats) {
+            def results
+
+            if (filtercat.platformProperty) {
+                results = BioAssayPlatform.createCriteria().list {
+                    projections {
+                        distinct(filtercat.platformProperty)
+                    }
+                    order(filtercat.platformProperty, "asc")
+                }
+            } else if (filtercat.prefix) {
+                results = ConceptCode.createCriteria().list {
+                    like("codeTypeName", filtercat.codeTypeName + ":%")
+                    order("codeName", "asc")
+                }
+            } else {
+                results = ConceptCode.createCriteria().list {
+                    eq("codeTypeName", filtercat.codeTypeName)
+                    order("codeName", "asc")
+                }
+            }
+
+            def choices = new TreeSet<Map>(new Comparator() {
+                int compare(Object o1, Object o2) {
+                    if (!o1 instanceof Map || !o2 instanceof Map) {
+                        return 0;
+                    }
+                    Map m1 = (Map) o1;
+                    Map m2 = (Map) o2;
+                    return o1.get('name').compareTo(o2.get('name'));
+                }
+            });
+            for (result in results) {
+                if (filtercat.platformProperty) {
+                    choices.add([name: result, uid: result])
+                } else if (filtercat.useText) {
+                    choices.add([name: result.codeName, uid: result.codeName])
+                } else {
+                    choices.add([name: result.codeName, uid: result.bioDataUid.uniqueId[0]])
+                }
+            }
+
+            if (choices) {
+                categories.add(["category": filtercat, "choices": choices])
+            }
+        }
+
+        return categories
+    }
+
+    /** Searches for all keywords for a given term (like %il%) */
+    def findSearchKeywords(category, term, max) {
+        log.info "Finding matches for ${term} in ${category}"
+
+        def user = AuthUser.findByUsername(springSecurityService.getPrincipal().username)
+
+        def c = SearchKeywordTerm.createCriteria()
+        def results = c.list {
+            if (term.size() > 0) {
+                like("keywordTerm", term.toUpperCase() + '%')
+            }
+
+            //TODO Special case for gene or SNP - rework to support multiple categories!
+            if ("GENE_OR_SNP".equals(category)) {
+                //or {
+                //	eq("dataCategory", "GENE")
+                //	like("keywordTerm", 'RS%')
+                //}
+                'in'("dataCategory", ["GENE", "SNP"])
+            } else if ("SNP".equals(category)) {
+                //like("keywordTerm", 'RS%')
+                eq("dataCategory", "SNP")
+            } else if ("ALL".compareToIgnoreCase(category) != 0) {
+                eq("dataCategory", category.toUpperCase())
+            }
+
+            if (!user.isAdmin()) {
+                log.info("User is not an admin so filter out gene lists or signatures that are not public")
+                or {
+                    isNull("ownerAuthUserId")
+                    eq("ownerAuthUserId", user.id)
+                }
+            }
+            maxResults(max)
+            order("rank", "asc")
+            order("termLength", "asc")
+            order("keywordTerm", "asc")
+        }
+        log.info("Search keywords found: " + results.size())
+
+        def keywords = []
+        def dupeList = []            // store category:keyword for a duplicate check until DB is cleaned up
+
+        for (result in results) {
+            def m = [:]
+            def sk = result
+            //////////////////////////////////////////////////////////////////////////////////
+            // HACK:  Duplicate check until DB is cleaned up
+            def dupeKey = sk.searchKeyword.displayDataCategory + ":" + sk.searchKeyword.keyword +
+                    ":" + sk.searchKeyword.bioDataId
+            if (dupeKey in dupeList) {
+                log.info "Found duplicate: " + dupeKey
+                continue
+            } else {
+                log.info "Found new entry, adding to the list: " + dupeList
+                dupeList << dupeKey
+            }
+            ///////////////////////////////////////////////////////////////////////////////////
+
+            m.put("label", sk.searchKeyword.keyword)
+            m.put("category", sk.searchKeyword.displayDataCategory)
+
+            //Further hack: Alter fields depending on the category
+            if (sk.searchKeyword.dataCategory.equals("DISEASE") || sk.searchKeyword.dataCategory.equals("OBSERVATION")) {
+                m.put("categoryId", sk.searchKeyword.dataCategory)
+                m.put("id", sk.searchKeyword.uniqueId)
+            } else {
+                m.put("categoryId", sk.searchKeyword.dataCategory)
+                m.put("id", sk.searchKeyword.id)
+            }
+            if ("TEXT".compareToIgnoreCase(sk.searchKeyword.dataCategory) != 0) {
+                def synonyms = org.transmart.biomart.BioDataExternalCode.findAllWhere(bioDataId: sk.searchKeyword.bioDataId, codeType: "SYNONYM")
+                def synList = new StringBuilder()
+                for (synonym in synonyms) {
+                    if (synList.size() > 0) {
+                        synList.append(", ")
+                    } else {
+                        synList.append("(")
+                    }
+                    synList.append(synonym.code)
+                }
+                if (synList.size() > 0) {
+                    synList.append(")")
+                }
+                m.put("synonyms", synList.toString())
+            }
+            keywords.add(m)
+        }
+
+        /*
+        * Get results from Bio Concept Code table
+        */
+
+        if (category.equals("ALL")) {
+            results = ConceptCode.createCriteria().list {
+                if (term.size() > 0) {
+                    like("bioConceptCode", term.toUpperCase().replace(" ", "_") + '%')
+                }
+                or {
+                    'in'("codeTypeName", filtercats*.codeTypeName)
+                }
+                maxResults(max)
+                order("bioConceptCode", "asc")
+            }
+            log.info("Bio concept code keywords found: " + results.size())
+
+            for (result in results) {
+                def m = [:]
+
+                //Get display name by category
+                def cat = filtercats.find { result.codeTypeName.startsWith(it.codeTypeName) }
+
+                m.put("label", result.codeName)
+                m.put("category", cat.displayName)
+                m.put("categoryId", cat.category)
+                if (cat.useText) {
+                    m.put("id", result.codeName)
+                } else {
+                    m.put("id", result.bioDataUid.uniqueId[0])
+                }
+                if (!keywords.find { it.id.equals(m.id) }) {
+                    keywords.add(m)
+                }
+
+            }
+
+            //If we're not over the maximum result threshold, query the platform table as well
+            if (keywords.size() < max) {
+
+                //Perform a query for each platform field
+                for (cat in filtercats) {
+                    if (cat.platformProperty) {
+                        results = BioAssayPlatform.createCriteria().list {
+                            ilike(cat.platformProperty, term + '%')
+                            maxResults(max)
+                            order(cat.platformProperty, "asc")
+                        }
+                        log.info("Platform " + cat.platformProperty + " keywords found: " + results.size())
+
+                        for (result in results) {
+                            def m = [:]
+
+                            m.put("label", result."${cat.platformProperty}")
+                            m.put("category", cat.displayName)
+                            m.put("categoryId", cat.category)
+                            m.put("id", result."${cat.platformProperty}")
+
+                            if (!keywords.find { it.id.equals(m.id) && it.category.equals(m.category) }) {
+                                keywords.add(m)
+                            }
+                        }
+                    }
+                }
+            }
+        }
 
 
-	
-	/** Searches for all keywords for a given term (like %il%) */
-	def findSearchKeywords(category, term)	{
-		log.info "Finding matches for ${term} in ${category}"
-		
-		def user = AuthUser.findByUsername(springSecurityService.getPrincipal().username)
-		
-		def c = SearchKeywordTerm.createCriteria()
-		def results = c.list 	{
-			if (term.size() > 0)	{
-				ilike("keywordTerm", '%' + term + '%')
-			}
-			if (category.class.name.toLowerCase() == 'java.util.arraylist') {
-				searchKeyword	{
-					inList("dataCategory", category)
-				}
-			}
-			else  {
-				if ("ALL".compareToIgnoreCase(category) != 0)	{
-					searchKeyword	{
-						eq("dataCategory", category, [ignoreCase: true])
-					}
-				}
-			}			
-			
-			if (!user.isAdmin())	{
-				log.info("User is not an admin so filter out gene lists or signatures that are not public")
-				or	{
-					isNull("ownerAuthUserId")
-					eq("ownerAuthUserId", user.id)
-				}
-			}
-			maxResults(20)
-			order("rank", "asc")
-			order("termLength", "asc")
-			order("keywordTerm", "asc")
-		}
-		log.info("Search keywords found: " + results.size())
-		
-		def keywords = []
-		def dupeList = []			// store category:keyword for a duplicate check until DB is cleaned up
-		
-		for (result in results)	{
-			def m = [:]
-			def sk = result
-			//////////////////////////////////////////////////////////////////////////////////
-			// HACK:  Duplicate check until DB is cleaned up
-			def dupeKey = sk.searchKeyword.displayDataCategory + ":" +sk.searchKeyword.keyword
-			if (dupeKey in dupeList)	{
-				log.info "Found duplicate: " + dupeKey
-				continue
-			} else	{
-				log.info "Found new entry, adding to the list: " + dupeList
-				dupeList << dupeKey
-			}
-			///////////////////////////////////////////////////////////////////////////////////
-			m.put("label", sk.searchKeyword.keyword)
-			m.put("id", , sk.searchKeyword.id)
-			m.put("category", sk.searchKeyword.displayDataCategory)
-			m.put("categoryId", sk.searchKeyword.dataCategory)
-			if ("TEXT".compareToIgnoreCase(sk.searchKeyword.dataCategory) != 0)	{
-				def synonyms = BioDataExternalCode.findAllWhere(bioDataId: sk.searchKeyword.bioDataId, codeType: "SYNONYM")
-				def synList = new StringBuilder()
-				for (synonym in synonyms)	{
-					if (synList.size() > 0)	{
-						synList.append(", ")
-					} else	{
-						synList.append("(")
-					}
-					synList.append(synonym.code)
-				}
-				if (synList.size() > 0)	{
-					synList.append(")")
-				}
-				m.put("synonyms", synList.toString())
-			}
-			keywords.add(m)
-		}
-		return keywords
-	 }
-	/**
-	 * This will render a UI where the user can pick an experiment from a list of all the experiments in the system. Selection of multiple studies is allowed.
-	 */
-	def browseAnalysisMultiSelect = {
-		def analyses = org.transmart.biomart.BioAssayAnalysis.executeQuery("select id, name, etlId from BioAssayAnalysis b order by b.name");
-	 	render(template:'browseMulti',model:[analyses:analyses])
-	}
+        return keywords
+    }
 
-	/**
-	 * convert pathways to a list of genes
-	 */
-	def expandPathwayToGenes(pathwayIds){
-		return expandPathwayToGenes(pathwayIds, null)
-	}
+    /** Searches for all keywords for a given term (like %il%) */
+    def findSearchKeywords(category, term) {
+        log.info "Finding matches for ${term} in ${category}"
 
-	/**
-	 * convert pathways to a list of genes
-	 */
-	def expandPathwayToGenes(pathwayIds, Long max){
-		if(pathwayIds==null || pathwayIds.length()==0){
-			return []
-		}
-		def query = "select DISTINCT k from org.transmart.searchapp.SearchKeyword k, org.transmart.biomart.BioDataCorrelation c where k.bioDataId=c.associatedBioDataId and c.bioDataId in (" + pathwayIds + ") ORDER BY k.keyword"
-		if(max!=null)
-			return SearchKeyword.executeQuery(query, [max:max])
-		else
-			return SearchKeyword.executeQuery(query)
+        def user = AuthUser.findByUsername(springSecurityService.getPrincipal().username)
+
+        def c = SearchKeywordTerm.createCriteria()
+        def results = c.list {
+            if (term.size() > 0) {
+                ilike("keywordTerm", '%' + term + '%')
+            }
+            if (category.class.name.toLowerCase() == 'java.util.arraylist') {
+                searchKeyword {
+                    inList("dataCategory", category)
+                }
+            } else {
+                if ("ALL".compareToIgnoreCase(category) != 0) {
+                    searchKeyword {
+                        eq("dataCategory", category, [ignoreCase: true])
+                    }
+                }
+            }
+
+            if (!user.isAdmin()) {
+                log.info("User is not an admin so filter out gene lists or signatures that are not public")
+                or {
+                    isNull("ownerAuthUserId")
+                    eq("ownerAuthUserId", user.id)
+                }
+            }
+            maxResults(20)
+            order("rank", "asc")
+            order("termLength", "asc")
+            order("keywordTerm", "asc")
+        }
+        log.info("Search keywords found: " + results.size())
+
+        def keywords = []
+        def dupeList = []            // store category:keyword for a duplicate check until DB is cleaned up
+
+        for (result in results) {
+            def m = [:]
+            def sk = result
+            //////////////////////////////////////////////////////////////////////////////////
+            // HACK:  Duplicate check until DB is cleaned up
+            def dupeKey = sk.searchKeyword.displayDataCategory + ":" + sk.searchKeyword.keyword
+            if (dupeKey in dupeList) {
+                log.info "Found duplicate: " + dupeKey
+                continue
+            } else {
+                log.info "Found new entry, adding to the list: " + dupeList
+                dupeList << dupeKey
+            }
+            ///////////////////////////////////////////////////////////////////////////////////
+            m.put("label", sk.searchKeyword.keyword)
+            m.put("id", sk.searchKeyword.id)
+            m.put("category", sk.searchKeyword.displayDataCategory)
+            m.put("categoryId", sk.searchKeyword.dataCategory)
+            if ("TEXT".compareToIgnoreCase(sk.searchKeyword.dataCategory) != 0) {
+                def synonyms = BioDataExternalCode.findAllWhere(bioDataId: sk.searchKeyword.bioDataId, codeType: "SYNONYM")
+                def synList = new StringBuilder()
+                for (synonym in synonyms) {
+                    if (synList.size() > 0) {
+                        synList.append(", ")
+                    } else {
+                        synList.append("(")
+                    }
+                    synList.append(synonym.code)
+                }
+                if (synList.size() > 0) {
+                    synList.append(")")
+                }
+                m.put("synonyms", synList.toString())
+            }
+            keywords.add(m)
+        }
+        return keywords
+    }
+    /**
+     * This will render a UI where the user can pick an experiment from a list of all the experiments in the system. Selection of multiple studies is allowed.
+     */
+    def browseAnalysisMultiSelect = {
+        def analyses = org.transmart.biomart.BioAssayAnalysis.executeQuery("select id, name, etlId from BioAssayAnalysis b order by b.name");
+        render(template: 'browseMulti', model: [analyses: analyses])
+    }
+
+    /**
+     * convert pathways to a list of genes
+     */
+    def expandPathwayToGenes(pathwayIds) {
+        return expandPathwayToGenes(pathwayIds, null)
+    }
+
+    /**
+     * convert pathways to a list of genes
+     */
+    def expandPathwayToGenes(pathwayIds, Long max) {
+        if (pathwayIds == null || pathwayIds.length() == 0) {
+            return []
+        }
+        def query = "select DISTINCT k from org.transmart.searchapp.SearchKeyword k, org.transmart.biomart.BioDataCorrelation c where k.bioDataId=c.associatedBioDataId and c.bioDataId in (" + pathwayIds + ") ORDER BY k.keyword"
+        if (max != null)
+            return SearchKeyword.executeQuery(query, [max: max])
+        else
+            return SearchKeyword.executeQuery(query)
 
 
-	}
+    }
 
-	def expandAllListToGenes(pathwayIds){
-		return expandAllListToGenes(pathwayIds, null);
-	}
+    def expandAllListToGenes(pathwayIds) {
+        return expandAllListToGenes(pathwayIds, null);
+    }
 
-	def expandAllListToGenes(pathwayIds, Long max){
-		if(pathwayIds==null || pathwayIds.length()==0){
-			return []
-		}
-		def result = [];
-		// find pathways
-		def query = "select DISTINCT k from org.transmart.searchapp.SearchKeyword k, org.transmart.biomart.BioDataCorrelation c where k.bioDataId=c.associatedBioDataId and c.bioDataId in (" + pathwayIds + ") ORDER BY k.keyword"
-		// find gene sigs
-		def query2 = "select DISTINCT k from org.transmart.searchapp.SearchKeyword k, org.transmart.searchapp.SearchBioMarkerCorrelFastMV c where k.bioDataId=c.assocBioMarkerId and c.domainObjectId in (" + pathwayIds + ") ORDER BY k.keyword"
-		if(max!=null)
-			result.addAll(SearchKeyword.executeQuery(query, [max:max]))
-			if(result.size()<max){
-				result.addAll(SearchKeyword.executeQuery(query2,[max:(max-result.size())]));
-			}
-		else{
-			result.addAll(SearchKeyword.executeQuery(query));
-			result.addAll(SearchKeyword.executeQuery(query2));
-		}
-			return result;
+    def expandAllListToGenes(pathwayIds, Long max) {
+        if (pathwayIds == null || pathwayIds.length() == 0) {
+            return []
+        }
+        def result = [];
+        // find pathways
+        def query = "select DISTINCT k from org.transmart.searchapp.SearchKeyword k, org.transmart.biomart.BioDataCorrelation c where k.bioDataId=c.associatedBioDataId and c.bioDataId in (" + pathwayIds + ") ORDER BY k.keyword"
+        // find gene sigs
+        def query2 = "select DISTINCT k from org.transmart.searchapp.SearchKeyword k, org.transmart.searchapp.SearchBioMarkerCorrelFastMV c where k.bioDataId=c.assocBioMarkerId and c.domainObjectId in (" + pathwayIds + ") ORDER BY k.keyword"
+        if (max != null)
+            result.addAll(SearchKeyword.executeQuery(query, [max: max]))
+        if (result.size() < max) {
+            result.addAll(SearchKeyword.executeQuery(query2, [max: (max - result.size())]));
+        } else {
+            result.addAll(SearchKeyword.executeQuery(query));
+            result.addAll(SearchKeyword.executeQuery(query2));
+        }
+        return result;
 
-	}
+    }
 
-	/**
-	 * link a GeneSignature new instance to search
-	 */
-	def newGeneSignatureLink(GeneSignature gs, boolean bFlush) {
-		// link  gs to search
-		SearchKeyword keyword = createSearchKeywordFromGeneSig(gs)
+    /**
+     * link a GeneSignature new instance to search
+     */
+    def newGeneSignatureLink(GeneSignature gs, boolean bFlush) {
+        // link  gs to search
+        SearchKeyword keyword = createSearchKeywordFromGeneSig(gs)
 
-		keyword.validate()
-		println("keyword validate()")
+        keyword.validate()
+        println("keyword validate()")
 
-		if(keyword.hasErrors()){
-			println("WARN: SearchKeyword validation error!")
-			keyword.errors.each { println it }
-		}
-		println("INFO: saving new SearchKeyword!")
-		keyword.save(flush: bFlush)
-	}
+        if (keyword.hasErrors()) {
+            println("WARN: SearchKeyword validation error!")
+            keyword.errors.each { println it }
+        }
+        println("INFO: saving new SearchKeyword!")
+        keyword.save(flush: bFlush)
+    }
 
-	/**
-	 * update GeneSignature/List link to search
-	 */
-	def updateGeneSignatureLink(GeneSignature gs, String domainKey, boolean bFlush) {
-		// find keyword record
-		SearchKeyword keyword = SearchKeyword.findByBioDataIdAndDataCategory(gs.id, domainKey)
-		println("INFO: retrieved "+keyword)
+    /**
+     * update GeneSignature/List link to search
+     */
+    def updateGeneSignatureLink(GeneSignature gs, String domainKey, boolean bFlush) {
+        // find keyword record
+        SearchKeyword keyword = SearchKeyword.findByBioDataIdAndDataCategory(gs.id, domainKey)
+        println("INFO: retrieved " + keyword)
 
-		// delete search keywords
-		if(gs.deletedFlag || (domainKey==GeneSignature.DOMAIN_KEY_GL && gs.foldChgMetricConceptCode.bioConceptCode=='NOT_USED')) {
-			if(keyword!=null) keyword.delete(flush: bFlush)
-		} else {
-			// add if does not exist
-			if(keyword==null) {
-				keyword = createSearchKeywordFromGeneSig(gs, domainKey)
-			} else {
-				// update keyword
-				keyword.keyword = gs.name
-				keyword.ownerAuthUserId = gs.publicFlag ? null : gs.createdByAuthUser.id
-				keyword.terms.each {
-					println("INFO: "+it)
-					it.keywordTerm = gs.name.toUpperCase()
-					it.ownerAuthUserId = gs.publicFlag ? null : gs.createdByAuthUser.id
-					//println("INFO: setting owner to: "+it.ownerAuthUserId)
-				}
-			}
+        // delete search keywords
+        if (gs.deletedFlag || (domainKey == GeneSignature.DOMAIN_KEY_GL && gs.foldChgMetricConceptCode.bioConceptCode != 'NOT_USED') || (domainKey == GeneSignature.DOMAIN_KEY && gs.foldChgMetricConceptCode.bioConceptCode == 'NOT_USED')) {
+            if (keyword != null) keyword.delete(flush: bFlush)
+        } else {
+            // add if does not exist
+            if (keyword == null) {
+                keyword = createSearchKeywordFromGeneSig(gs, domainKey)
+            } else {
+                // update keyword
+                keyword.keyword = gs.name
+                keyword.ownerAuthUserId = gs.publicFlag ? null : gs.createdByAuthUser.id
+                keyword.terms.each {
+                    println("INFO: " + it)
+                    it.keywordTerm = gs.name.toUpperCase()
+                    it.ownerAuthUserId = gs.publicFlag ? null : gs.createdByAuthUser.id
+                    //println("INFO: setting owner to: "+it.ownerAuthUserId)
+                }
+            }
 
-			keyword.validate()
-			if(keyword.hasErrors()){
-				println("WARN: SearchKeyword validation error!")
-				keyword.errors.each { println it }
-			}
-			println("INFO: trying to save SearchKeyword")
-			keyword.save(flush: bFlush)
-		}
-	}
+            keyword.validate()
+            if (keyword.hasErrors()) {
+                println("WARN: SearchKeyword validation error!")
+                keyword.errors.each { println it }
+            }
+            println("INFO: trying to save SearchKeyword")
+            keyword.save(flush: bFlush)
+        }
+    }
 
-	/**
-	 * create a new SearchKeyord for a GeneSignatute
-	 */
-	def createSearchKeywordFromGeneSig(GeneSignature gs, String domainKey) {
-		println("INFO: creating SearchKeyword for GS: "+gs.name+ "["+domainKey+"]")
+    /**
+     * create a new SearchKeyord for a GeneSignatute
+     */
+    def createSearchKeywordFromGeneSig(GeneSignature gs, String domainKey) {
+        println("INFO: creating SearchKeyword for GS: " + gs.name + "[" + domainKey + "]")
 
-		// display category GS or GL?
-		def displayName = (domainKey==GeneSignature.DOMAIN_KEY) ? GeneSignature.DISPLAY_TAG : GeneSignature.DISPLAY_TAG_GL
-		def uniqueId = domainKey+":"+gs.id
+        // display category GS or GL?
+        def displayName = (domainKey == GeneSignature.DOMAIN_KEY) ? GeneSignature.DISPLAY_TAG : GeneSignature.DISPLAY_TAG_GL
+        def uniqueId = domainKey + ":" + gs.id
 
-		SearchKeyword keyword = new SearchKeyword()
-		keyword.properties.keyword = gs.name
-		keyword.properties.bioDataId = gs.id
-		keyword.properties.uniqueId = uniqueId
-		keyword.properties.dataCategory = domainKey
-		keyword.properties.displayDataCategory = displayName
-		keyword.properties.dataSource="Internal"
-		if(!gs.publicFlag) keyword.properties.ownerAuthUserId = gs.createdByAuthUser?.id
+        SearchKeyword keyword = new SearchKeyword()
+        keyword.properties.keyword = gs.name
+        keyword.properties.bioDataId = gs.id
+        keyword.properties.uniqueId = uniqueId
+        keyword.properties.dataCategory = domainKey
+        keyword.properties.displayDataCategory = displayName
+        keyword.properties.dataSource = "Internal"
+        if (!gs.publicFlag) keyword.properties.ownerAuthUserId = gs.createdByAuthUser?.id
 
-		// keyword term
-		SearchKeywordTerm term = new SearchKeywordTerm()
-		term.properties.keywordTerm = gs.name.toUpperCase()
-		term.properties.rank = 1
-		term.properties.termLength = gs.name.length()
-		if(!gs.publicFlag) term.properties.ownerAuthUserId = gs.createdByAuthUser?.id
+        // keyword term
+        SearchKeywordTerm term = new SearchKeywordTerm()
+        term.properties.keywordTerm = gs.name.toUpperCase()
+        term.properties.rank = 1
+        term.properties.termLength = gs.name.length()
+        if (!gs.publicFlag) term.properties.ownerAuthUserId = gs.createdByAuthUser?.id
 
-		// associate term
-		keyword.addToTerms(term)
+        // associate term
+        keyword.addToTerms(term)
 
-		//println(keyword)
-		//println("properties:\n"+keyword.properties)
-		return keyword;
-	}
+        //println(keyword)
+        //println("properties:\n"+keyword.properties)
+        return keyword;
+    }
 
 }
